get sorted indicators and results from the grid
